Trying to get startServer script up and running. <Jaideep>

+++ start-server.sh
@@ -1,3 +1,6 @@
+if [ -f .bash_profile ]; then
 cat as_pid_file | xargs kill -9
-nohup java -jar platform-1.0-SNAPSHOT.jar server $1 > server.log &
+else
+java -jar platform-1.0-SNAPSHOT.jar server $1 > server.log &
 echo $! > as_pid_file
+fi
